rag-retrievers/requirements.txt
transformers
langchain
Llama-index-core
streamlit
nebula3-python
python-dotenv
matplotlib
torch
sentence-transformers
sentencepiece

llama-index-embeddings-huggingface
llama-index-embeddings-langchain
llama-index-graph-stores-nebula
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-agent-openai

lxml 
unstructured 
oss2
markdown2
markdownify